Separated tests with a base test and a report for each page tested
